Delete Selected Message

+++ at_chat_flutter/example/lib/third_screen.dart
@@ -17,8 +17,17 @@ class _ThirdScreenState extends State<ThirdScreen> {
             icon: const Icon(Icons.delete),
             tooltip: 'Show Snackbar',
             onPressed: () async {
-              bool result = await deleteMessages();
-              String message = result ? 
+              final confirmed = await confirm(
+                this.context,
+                title: 'Confirm Dialog',
+                message: 'Do you want to delete this chat?',
+                positiveActionTitle: 'Yes',
+                negativeActionTitle: 'No',
+              );
+              if (!confirmed) return;
+
+              var result = await deleteMessages();
+              var message = result ?
                 'Messages are deleted' : 'Failed to delete';
               ScaffoldMessenger.of(context).showSnackBar(
                  SnackBar(content: Text(message)));
